move controllers from controllers to controller dir; rename controllers to PascalCase
