+++ src/main/java/com/booking/controller/GetFinancialReportController.java
@@ -10,18 +10,14 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.Collection;
 
 @RestController
-@RequestMapping("/api/getFavoriteAccommodations")
-public class getFavoriteAccommodationsController {
+@RequestMapping("/api/getNotifications")
+public class GetFinancialReportController {
     // TODO: implement
     @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<Collection<Boolean>> getFavoriteAccommodations() {
-        // Collection<Boolean> favoriteAccommodations = favoriteAccommodationService.findAll();
+    public ResponseEntity<Collection<Boolean>> getFinancialReport() {
+        // FinancialReport financialReport = someService.getFinancialReport();
 
-        // Boolean success = reviewService.delete(data);
-        // if (success) {
-        // return new ResponseEntity<Collection<Boolean>>(favoriteAccommodations, HttpStatus.OK);
-        // }
-        // return a 4XX status
+        // return new ResponseEntity<Collection<Boolean>>(financialReport, HttpStatus.OK);
 
         return new ResponseEntity<Collection<Boolean>>(HttpStatus.OK);
     }
