+++ src/main/java/com/booking/controller/GetNotificationsController.java
@@ -9,7 +9,7 @@ import java.util.Collection;
 
 @RestController
 @RequestMapping("/api/getNotifications")
-public class getNotificationsController {
+public class GetNotificationsController {
     // TODO: implement
     @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Collection<Boolean>> getNotifications() {
